requirements.txt
cryptography==44.0.1
numpy==1.26.4  # Updated for compatibility
pandas==2.2.3
Pillow==10.0.1  # Updated for compatibility
plotly==6.0.0
scikit_learn==1.3.0  # Updated for compatibility
streamlit==1.22.0 # Streamlit 1.41.1 may have issues
streamlit_option_menu==0.4.0
tensorflow==2.13.0 # Consider a lower TF Version